Ensure consistent deadlines between simulator and Spark (hack)
We found that deadlines for task graphs weren't consistent between the
simulator and Spark even with the same RNG seed being used, due to the
fact that EventTime keeps a global RNG it uses for all of its fuzzing
and both deadlines and runtime variances are fuzzed.  Since in
simulator runs, task deadlines are all calculated at the start and
runtime variances are calculated later, and in Spark, task deadlines
and runtime variances are calculated throughout the experiment
lifecycle, different deadline variances are obtained between simulator
and Spark runs on the same experiment.

Our solution is to pass a unique RNG used just for calculating
deadline variances to the fuzzer.  This RNG is hardcoded with a seed
of 42; this is fine for experiments but it should probably be changed
to the random_seed command line flag.