Spark service: Make job graph names the same as in TpchLoader
Originally, the service named job graphs in the form
Q<query>[<spark-app-id>], where Spark sets the app id to
app-<timestamp>-<index>, while TPC-H data loader named job graphs in
the form Q<query>[<index>].  This commit changes how the service names
job graphs by passing the index as an argument to the TpchQuery Spark
application, which will then be forwarded into the Servicer through
RegisterTaskGraph as a part of the query name.  RegisterTaskGraph then
uses the index to name the job graph.

This ensures that the job graph names are always the same between a
Spark run and a simulator run, irrespective of when the task graphs
are actually released during a Spark run (which can be
nondeterministic).  The intent is to use these names to generate
deadlines for the task graphs, so that deadlines are always consistent
between Spark and simulator runs.

This change requires a corresponding change to tpch-spark to forward
the index to the Servicer.