TetriSchedScheduler: Reconsider tasks that couldn't be placed
To avoid needless reruns, TetriSchedScheduler does not run the
scheduler if there are no tasks which are not scheduled, not part of a
task graph that has been previously considered, and not part of a task
graph that has been cancelled.  We remove this second condition to
account for situations in which a task graph is considered and its
tasks scheduled, but the tasks failed to be placed (for instance, if
another task on the same worker finished late, taking up resources).
In such cases, the task graph would not be cancelled and might still
be able to be completed, so we need to run the scheduler again to try
to schedule the tasks that could not be placed before.